medease-agent/requirements.txt
twilio
fastapi
uvicorn[standard]
google-cloud-bigquery
google-adk
pandas
pydantic
httpx